// resources/js/api/generated/models/getApiV1ServersServeridStacks200StacksItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1ServersServeridStacks200StacksItem = {
  compose_file: string;
  is_healthy: boolean;
  name: string;
  path: string;
  running_containers: number;
  /** @minimum 0 */
  server_id: number;
  server_name: string;
  total_containers: number;
};